+++ agent_bitcoin/l402/__init__.py
@@ -1,0 +1,10 @@
+"""L402 (Lightning HTTP 402) client helpers."""
+
+from .client import L402Challenge, L402Client, L402Response, parse_www_authenticate
+
+__all__ = [
+    "L402Challenge",
+    "L402Client",
+    "L402Response",
+    "parse_www_authenticate",
+]